+++ src/client/failure/event-adapter.test.ts
@@ -11,6 +11,7 @@ import { getLocal } from 'mockttp';
 import { SuperJson } from '../../internal';
 import { bindResponseError } from '../../internal/errors.helpers';
 import { ok, sleep } from '../../lib';
+import { ServiceSelector } from '../../lib/services';
 import {
   createTypedClient,
   invalidateSuperfaceClientCache,
@@ -436,21 +437,21 @@ function spyOnCacheBoundProfileProvider(client: SuperfaceClientBase) {
     firstMockProfileDocument,
     firstMockMapDocument,
     'provider',
-    { baseUrl: mockServer.url, security: [] },
+    { services: ServiceSelector.withDefaultUrl(mockServer.url), security: [] },
     client
   );
   const secondMockBoundProfileProvider = new BoundProfileProvider(
     firstMockProfileDocument,
     secondMockMapDocument,
     'second',
-    { baseUrl: mockServer.url, security: [] },
+    { services: ServiceSelector.withDefaultUrl(mockServer.url), security: [] },
     client
   );
   const thirdMockBoundProfileProvider = new BoundProfileProvider(
     secondMockProfiledDocument,
     thirdMockMapDocument,
     'third',
-    { baseUrl: mockServer.url, security: [] },
+    { services: ServiceSelector.withDefaultUrl(mockServer.url), security: [] },
     client
   );
   const cacheBoundProfileProviderSpy = jest