+++ src/client/profile-provider.ts
@@ -30,7 +30,6 @@ import {
   providersDoNotMatchError,
   referencedFileNotFoundError,
   securityNotFoundError,
-  serviceNotFoundError,
 } from '../internal/errors.helpers';
 import {
   MapInterpreter,
@@ -51,6 +50,7 @@ import { mergeSecurity } from '../internal/superjson/mutate';
 import { err, ok, Result } from '../lib';
 import { Events, Interceptable } from '../lib/events';
 import { CrossFetch } from '../lib/fetch';
+import { IServiceSelector, ServiceSelector } from '../lib/services';
 import { MapInterpreterEventAdapter } from './failure/map-interpreter-adapter';
 import { ProfileConfiguration } from './profile';
 import { ProviderConfiguration } from './provider';
@@ -67,18 +67,19 @@ function profileAstId(ast: ProfileDocumentNode): string {
 const boundProfileProviderDebug = createDebug(
   'superface:bound-profile-provider'
 );
+export type AuthCache = { digest?: Record<string, string> };
 
 export class BoundProfileProvider {
-  //TODO: Interceptable and set metadata
+  // TODO: Interceptable and set metadata
   private profileValidator: ProfileParameterValidator;
-  private fetchInstance: FetchInstance & Interceptable;
+  private fetchInstance: FetchInstance & Interceptable & AuthCache;
 
   constructor(
     private readonly profileAst: ProfileDocumentNode,
     private readonly mapAst: MapDocumentNode,
     private readonly providerName: string,
     readonly configuration: {
-      baseUrl: string;
+      services: IServiceSelector;
       profileProviderSettings?: NormalizedProfileProviderSettings;
       security: SecurityConfiguration[];
       parameters?: Record<string, string>;
@@ -124,7 +125,8 @@ export class BoundProfileProvider {
     TResult = any
   >(
     usecase: string,
-    input?: TInput
+    input?: TInput,
+    parameters?: Record<string, string>
   ): Promise<Result<TResult, ProfileParameterError | MapInterpreterError>> {
     this.fetchInstance.metadata = {
       profile: profileAstId(this.profileAst),
@@ -149,9 +151,12 @@ export class BoundProfileProvider {
       {
         input: composedInput,
         usecase,
-        serviceBaseUrl: this.configuration.baseUrl,
+        services: this.configuration.services,
         security: this.configuration.security,
-        parameters: this.configuration.parameters,
+        parameters: this.mergeParameters(
+          parameters,
+          this.configuration.parameters
+        ),
       },
       {
         fetchInstance: this.fetchInstance,
@@ -180,6 +185,24 @@ export class BoundProfileProvider {
     forceCast<TResult>(result.value);
 
     return ok(result.value);
+  }
+
+  private mergeParameters(
+    parameters?: Record<string, string>,
+    providerParameters?: Record<string, string>
+  ): Record<string, string> | undefined {
+    if (parameters === undefined) {
+      return providerParameters;
+    }
+
+    if (providerParameters === undefined) {
+      return parameters;
+    }
+
+    return {
+      ...providerParameters,
+      ...parameters,
+    };
   }
 }
 
@@ -310,19 +333,6 @@ export class ProfileProvider {
       );
     }
 
-    // prepare service info
-    const serviceId = providerInfo.defaultService;
-    const baseUrl = providerInfo.services.find(
-      s => s.id === serviceId
-    )?.baseUrl;
-    if (baseUrl === undefined) {
-      throw serviceNotFoundError(
-        serviceId,
-        providerName,
-        serviceId === providerInfo.defaultService
-      );
-    }
-
     const securityConfiguration = this.resolveSecurityConfiguration(
       providerInfo.securitySchemes ?? [],
       securityValues,
@@ -334,15 +344,18 @@ export class ProfileProvider {
       mapAst,
       providerInfo.name,
       {
-        baseUrl,
+        services: new ServiceSelector(
+          providerInfo.services,
+          providerInfo.defaultService
+        ),
         profileProviderSettings:
           this.superJson.normalized.profiles[profileId]?.providers[
             providerInfo.name
           ],
         security: securityConfiguration,
         parameters: this.resolveIntegrationParameters(
-          this.superJson.normalized.providers[providerInfo.name]?.parameters,
-          providerInfo
+          providerInfo,
+          this.superJson.normalized.providers[providerInfo.name]?.parameters
         ),
       },
       this.events
@@ -350,8 +363,8 @@ export class ProfileProvider {
   }
 
   private resolveIntegrationParameters(
-    superJsonParameters: { [key: string]: string } | undefined,
-    providerJson: ProviderJson
+    providerJson: ProviderJson,
+    superJsonParameters?: Record<string, string>
   ): Record<string, string> | undefined {
     if (superJsonParameters === undefined) {
       return undefined;
@@ -363,7 +376,7 @@ export class ProfileProvider {
       providerJsonParameters.length === 0
     ) {
       console.warn(
-        `Warning: Super.json defines integration parameters but provider.json does not`
+        'Warning: Super.json defines integration parameters but provider.json does not'
       );
     }
     const result: Record<string, string> = {};
@@ -372,12 +385,13 @@ export class ProfileProvider {
       providerJson.name,
       providerJsonParameters
     );
-    //Resolve parameters defined in super.json
+
+    // Resolve parameters defined in super.json
     for (const [key, value] of Object.entries(superJsonParameters)) {
       const providerJsonParameter = providerJsonParameters.find(
         parameter => parameter.name === key
       );
-      //If value name and prepared value equals we are dealing with unset env
+      // If value name and prepared value equals we are dealing with unset env
       if (
         providerJsonParameter &&
         preparedParameters[providerJsonParameter.name] === value
@@ -387,13 +401,13 @@ export class ProfileProvider {
         }
       }
 
-      //Use original value
+      // Use original value
       if (!result[key]) {
         result[key] = value;
       }
     }
 
-    //Resolve parameters which are missing in super.json and have default value
+    // Resolve parameters which are missing in super.json and have default value
     for (const parameter of providerJsonParameters) {
       if (result[parameter.name] === undefined && parameter.default) {
         result[parameter.name] = parameter.default;