+++ src/internal/interpreter/interfaces.ts
@@ -1,4 +1,5 @@
 import {
+  ComlinkAssignmentNode,
   ComlinkListLiteralNode,
   ComlinkObjectLiteralNode,
   ComlinkPrimitiveLiteralNode,
@@ -37,6 +38,10 @@ export interface ProfileVisitor {
   ): unknown;
   visitComlinkPrimitiveLiteralNode(
     node: ComlinkPrimitiveLiteralNode,
+    ...parameters: unknown[]
+  ): unknown;
+  visitComlinkAssignmentNode(
+    node: ComlinkAssignmentNode,
     ...parameters: unknown[]
   ): unknown;
   visitEnumDefinitionNode(