+++ CHANGELOG.md
@@ -7,6 +7,32 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added support for service selection in http calls
+- Added support for digest authentication in http calls
+
+## [1.2.1] - 2022-01-19
+
+This is a recovery release after introduction of compatibility-breaking changes in version 1.2.0.
+
+- No changes since 1.1.0
+
+## [1.1.0] - 2021-12-22
+### Added
+- Pass integration parameters in `perform`
+
+### Fixed
+- Display query parameters in URL with `DEBUG=superface:http:sensitive`
+
+## [1.0.1] - 2021-11-24
+### Added
+- Use [JSON Pointer](https://datatracker.ietf.org/doc/html/rfc6901) (arrays not supported) to specify api token location in body
+
+### Fixed
+- ProfileParameterValidator now resolves named models before field references
+- Buffer serialization in request body
+- `undefined` values are removed when stringifying records
+
 ## [1.0.0] - 2021-11-04
 ### Added
 - Module for mocking of Superface components
@@ -261,7 +287,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Profile parameter validator
 - CI/CD workflows
 
-[Unreleased]: https://github.com/superfaceai/one-sdk-js/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/superfaceai/one-sdk-js/compare/v1.2.1...HEAD
+[1.2.1]: https://github.com/superfaceai/one-sdk-js/compare/v1.1.0...v1.2.1
+[1.1.0]: https://github.com/superfaceai/one-sdk-js/compare/v1.0.1...v1.1.0
+[1.0.1]: https://github.com/superfaceai/one-sdk-js/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/superfaceai/one-sdk-js/compare/v0.0.40...v1.0.0
 [0.0.40]: https://github.com/superfaceai/one-sdk-js/compare/v0.0.39...v0.0.40
 [0.0.39]: https://github.com/superfaceai/one-sdk-js/compare/v0.0.38...v0.0.39